// packages/protocol/typescript/generated/yiru/runtime/v1/git_status_pb.d.ts
export declare type GitWorkingStatus = Message<"yiru.runtime.v1.GitWorkingStatus"> & {
  /**
   * @generated from field: repeated yiru.runtime.v1.GitStatusEntry entries = 1;
   */
  entries: GitStatusEntry[];

  /**
   * @generated from field: yiru.runtime.v1.GitConflictOperation conflict_operation = 2;
   */
  conflictOperation: GitConflictOperation;

  /**
   * @generated from field: optional yiru.runtime.v1.GitUpstreamStatus upstream_status = 3;
   */
  upstreamStatus?: GitUpstreamStatus | undefined;

  /**
   * @generated from field: optional string head = 4;
   */
  head?: string | undefined;

  /**
   * @generated from field: optional string branch = 5;
   */
  branch?: string | undefined;

  /**
   * @generated from field: repeated string ignored_paths = 6;
   */
  ignoredPaths: string[];

  /**
   * @generated from field: optional bool did_hit_limit = 7;
   */
  didHitLimit?: boolean | undefined;

  /**
   * @generated from field: optional uint64 status_length = 8;
   */
  statusLength?: bigint | undefined;
};